+++ themes/book/app/routes/$.tsx
@@ -28,12 +28,15 @@ import {
 } from '@myst-theme/providers';
 import { MadeWithMyst } from '@myst-theme/icons';
 import { ComputeOptionsProvider, ThebeLoaderAndServer } from '@myst-theme/jupyter';
-import { ArticlePage } from '../components/ArticlePage.js';
-import type { TemplateOptions } from '../types.js';
+import { ArticlePage } from '~/components/ArticlePage';
+import type { TemplateOptions } from '~/types';
 import { useRouteError, isRouteErrorResponse } from '@remix-run/react';
-import { Footer } from '@myst-theme/site/dist/components/Footer';
-import TextWithLink from '@myst-theme/site/dist/components/TextWithLink';
-import ButtonWithIcon, { iconPositions } from '@myst-theme/site/dist/components/ButtonWithIcon';
+import { Footer } from '@myst-theme/site/dist/components/Custom/Footer';
+import { TextWithLink } from '@myst-theme/site/dist/components/Custom/TextWithLink';
+import {
+  ButtonWithIcon,
+  iconPositions,
+} from '@myst-theme/site/dist/components/Custom/ButtonWithIcon';
 import { EnvelopeIcon } from '@heroicons/react/24/solid';
 type ManifestProject = Required<SiteManifest>['projects'][0];
